server/requirements.txt
psycopg2-binary==2.8.4
environs==7.2.0
Flask==1.1.1
Flask-Bootstrap==3.3.7.1
APScheduler==3.6.3
requests==2.23.0
beautifulsoup4==4.8.2
connexion[swagger-ui]==2.7.0